refactor(composer): move the project picker into the composer header
On the empty-chat hero the picker floated in the gap between the greeting
and the composer card, belonging to neither surface. Astryx ChatComposer
has a header row above the input that nothing used; the picker goes
there, and the row is passed only while the chat is a draft, so it still
leaves the screen the moment the first message creates the session.

This keeps what moving it off the footer bought — the footer's other
controls persist for the whole session, and a chip that vanishes on send
broke that row's contract — without floating the control on the canvas.

The menu now opens upward, into the canvas the draft state leaves empty:
downward it covered the input the user is about to type in, and on a
short window it ran past the frame. The palette seam and the pinned-footer
rules follow the picker to a wrapper class in composer.css.

Verified in Storybook: the header row renders in the draft story and is
absent from every active-session story, the menu opens above the trigger
and stays inside the frame, and the pinned actions still hold.
