+++ apps/desktop/src/renderer/styles/hero.css
@@ -98,82 +98,3 @@
 .maka-hero-empty-chat:not(.maka-hero-deep-research) h1 {
     font: var(--maka-text-heading-1);
   }
-
-/* The workspace picker's row on the empty-chat canvas. It sits between the
-   greeting and the composer because the project is a session-creation
-   parameter — chosen once, before the first message, and gone with this hero
-   the moment the session exists. See packages/ui/src/workspace-picker.tsx. */
-.maka-hero-workspace {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: var(--space-1-5);
-}
-
-.maka-workspace-picker {
-  -webkit-app-region: no-drag;
-}
-
-.maka-workspace-picker-option {
-  min-width: 0;
-}
-
-/* Pin "add project" / "no project" to the bottom of the open menu so only the
-   project catalogue scrolls. Astryx `Selector` scrolls its whole list as one
-   region and exposes no footer slot, so the two rows are stuck in place here;
-   `workspace-picker.tsx` marks them with `-pinned`. Scoped to the hero because
-   the Layer renders the menu next to the trigger, not portaled to the body.
-
-   The opaque backdrop is painted by the list's `::after`, not by the rows: the
-   rows' own background carries Astryx's hover and keyboard-highlight tints, and
-   an opaque colour on them would win the cascade (this file's `components`
-   layer sits after `astryx-components`) and silently kill both. As a pinned
-   panel behind them it hides the scrolled rows without touching those states. */
-.maka-hero-workspace [role="listbox"] {
-  /* Astryx's `sm` option height and the list's own padding. Offsets stack from
-     the bottom edge in these units; if either changes upstream the footer
-     drifts, which is visible the moment the menu is opened. */
-  --maka-picker-row: 28px;
-  --maka-picker-pad: 4px;
-}
-
-.maka-hero-workspace [role="listbox"] > [role="option"]:has(.maka-workspace-picker-option-pinned),
-.maka-hero-workspace [role="listbox"] > [role="separator"]:has(+ [role="option"] + [role="option"]:last-child) {
-  position: sticky;
-  z-index: 1;
-}
-
-.maka-hero-workspace [role="listbox"] > [role="option"]:last-child:has(.maka-workspace-picker-option-pinned) {
-  bottom: var(--maka-picker-pad);
-}
-
-.maka-hero-workspace
-  [role="listbox"]
-  > [role="option"]:has(.maka-workspace-picker-option-pinned):has(+ [role="option"]:last-child) {
-  bottom: calc(var(--maka-picker-pad) + var(--maka-picker-row));
-}
-
-.maka-hero-workspace [role="listbox"] > [role="separator"]:has(+ [role="option"] + [role="option"]:last-child) {
-  bottom: calc(var(--maka-picker-pad) + 2 * var(--maka-picker-row));
-}
-
-/* Last in flow, pulled back over the footer rows and pinned with them. Its own
-   height would otherwise lengthen the scroll, so the negative margins cancel
-   it out. Only present while the footer is: a search that filters both actions
-   out leaves an ordinary scrolling list. */
-.maka-hero-workspace [role="listbox"]:has(.maka-workspace-picker-option-pinned)::after {
-  content: "";
-  display: block;
-  position: sticky;
-  bottom: 0;
-  height: calc(var(--maka-picker-pad) + 2 * var(--maka-picker-row) + 1px);
-  margin-top: calc(-2 * var(--maka-picker-row) - 1px);
-  margin-bottom: calc(-1 * var(--maka-picker-pad));
-  background: var(--color-background-popover);
-}
-
-.maka-workspace-picker-status {
-  font: var(--maka-text-supporting);
-  margin-left: auto;
-  color: var(--muted-foreground);
-}
