+++ apps/desktop/stories/app-shell.stories.tsx
@@ -158,19 +158,6 @@ const baseChatProps: ChatViewProps = {
   userLabel: '你',
   onNew: noop,
   onPromptSuggestion: noop,
-  // Production wires this whenever the shell has a project catalog
-  // (app-shell.tsx). It only reaches the screen in the draft state, so the
-  // active-session stories below carry it without rendering it.
-  workspacePicker: {
-    label: 'maka-agent',
-    branch: 'opencode/storybook-surface-coverage',
-    projects: catalogProjects.filter((item) => item.archivedAt === undefined),
-    selectedProjectId: 'project-maka',
-    onAdd: noop,
-    onSelectProject: noop,
-    onRelink: noop,
-    onSelectNoProject: noop,
-  },
 };
 
 const baseComposerProps: ComposerProps = {
@@ -470,6 +457,19 @@ export const NewChatComposer: Story = {
         newChatModel: { llmConnectionSlug: 'anthropic-main', model: 'claude-sonnet-4-5' },
         onPickNewChatModel: noop,
         onOpenModelSettings: noop,
+        // Draft-state only: app-shell.tsx passes this while no session exists,
+        // and the composer's header row exists only while it is passed. The
+        // active-session stories leave it out for the same reason.
+        workspacePicker: {
+          label: 'maka-agent',
+          branch: 'opencode/storybook-surface-coverage',
+          projects: catalogProjects.filter((item) => item.archivedAt === undefined),
+          selectedProjectId: 'project-maka',
+          onAdd: noop,
+          onSelectProject: noop,
+          onRelink: noop,
+          onSelectNoProject: noop,
+        },
       }}
     />
   ),
